+++ PingUI/ServiceModels/IConfiguration.cs
@@ -1,4 +1,6 @@
 using System.Collections.ObjectModel;
+using Avalonia;
+using PingUI.Interop;
 using PingUI.Models;
 
 namespace PingUI.ServiceModels;
@@ -36,6 +38,16 @@ public interface IConfiguration
 	/// </summary>
 	/// <value><see langword="null" /> for unknown, <see langword="true" /> to check for updates, <see langword="false" /> not to check for updates.</value>
 	bool? CheckOnlineForUpdates
+	{
+		get;
+		set;
+	}
+
+	/// <summary>
+	/// Indicates whether the app should remember and restore the window location on startup.
+	/// </summary>
+	/// <value>A <see cref="WindowPlacement" /> value to save and restore; otherwise <see langword="null" /> to not save and restore.</value>
+	WindowPlacement.WindowPlacementRecord? WindowBounds
 	{
 		get;
 		set;